Move index.html outside of the recipes folder
Here I have moved the index.html up one level, back directly to the odin-recipes folder, since having it inside of the recipes folder was a mistake in the first place, so I'm rectifying it here.

I'm not sure if it would be better to go back in time and fixing the commit where I created the index.html file inside of the recipes folder (neither do I remember how to do it in the first place), so this is the approach I'm taking.